+++ app/components/Navigation.tsx
@@ -98,7 +98,7 @@ const styles = StyleSheet.create({
   },
   logoImage: {
     height: '100%',
-    width: 200,
+    width: 250,
   },
   nav: {
     flexDirection: 'row',